change filenames to PSR-0/4, and change namespace to ybourque
incidentally strip accidental trailing whitespace
