+++ lib/GenderParse.php
@@ -1,4 +1,6 @@
 <?php
+namespace ybourque\Wikparser\lib;
+
 /***********************************************************************************/
 // This class is used to parse the wikitionary raw data in order to extract gender
 // for a given word.
@@ -11,7 +13,7 @@ class GenderParse {
 /***********************************************************************************/
 	private $langCode;			// language code (e.g. en, fr, da, etc.)
 	private $genderPattern;		// Gender regex pattern, set in config file
-	
+
 /***********************************************************************************/
 // construct
 /***********************************************************************************/
@@ -26,13 +28,13 @@ class GenderParse {
 	}
 /***********************************************************************************/
 // public methods; used to retrieve contents of variables
-/***********************************************************************************/	
+/***********************************************************************************/
 	public function get_gender($wikitext, $count) {
 		$genderArray = $this->extract_gender($wikitext, $count);
-		
+
 		include "./classes/class.strip.tags.php";
 		$stripTagsObject = new StripTags();
-		
+
 		return $stripTagsObject->strip_tags($genderArray, $this->langCode);
 	}
 /***********************************************************************************/
@@ -42,24 +44,24 @@ class GenderParse {
 /***********************************************************************************/
 	private function extract_gender($wikitext, $count) {
 		$tempGenderResults = array();
-		
+
 		if ($this->genderPattern != "") {
-			
+
 			preg_match_all($this->genderPattern, $wikitext, $matches);
-			
+
 			if (empty($matches[0]) !== true) {
 				foreach ($matches[0] as $value) {
 					$tempGenderResults[] = $value;
 				}
-			// Remove values based on the count provide by the user.	
+			// Remove values based on the count provide by the user.
 				$tempGenderResults = array_slice($tempGenderResults, 0, $count);
 			}
 		}
 		else {
 			die("No gender pattern specified for this language.");
 		}
-		
-	// Return results if array not empty.	
+
+	// Return results if array not empty.
 		if (empty($tempGenderResults) !==true) {
 			return array_unique($tempGenderResults);
 		}
